Replace the MassFlowSource_T.co on the "high" AFN ports with an Orifice.mo + Boundary_pT.mo. The pressure on the high side was set at 99 kPa, so that flow will still be driven into the "low" openings and out of the "high" openings.
Simulations and produces expected pressures and flow rates.  However, this testing revealed
that WeaBus does not contain the atmospheric pressure from the weather file.  Rather, it is
always 101325 Pa.  Looking through the WeaBus.mo code I realize that WeaDatBus.mo contains
Patm, but this is not placed on WeaBus.  This will have to be addressed in the future.